Bugfix for play-copy-assets
Pass the correct scoped resourceManaged to the resourceGenerator. There
is a difference in recent SBTs between the default resourceManaged
property and the one in compile scope. A slug representing the scope is
added to the directory path.

If this slug is not taken into account, the generated version.txt file
is placed at the wrong height in the directory structure. SBT still
manages it as a managed resource but there are problems later trying
to rebase the directory prefix because there is only a rebase match for
the directory prefix with the scope slug.

Anyway, a resourceGenerator in the compile scope should use the settings
for that scope. I presume the problem here is the resourceManaged on the
right hand side is bound to the default resourceManaged at the time
this resourceGenerator is added and not dependent on the scope we are in
at the time the resourceGenerator is run. Maybe?